Refactored dashboard to use dashboard controller
